move admin drawer into own chunk
