+++ frontend/src/components/admin/drawer/admin-menu.vue
@@ -34,7 +34,7 @@ import {
 } from "@mdi/js";
 import { mapActions, mapGetters, mapWritableState } from "pinia";
 
-import AdminStatusList from "@/components/admin/status-list.vue";
+import AdminStatusList from "@/components/admin/drawer/status-list.vue";
 import { useAdminStore } from "@/stores/admin";
 import { useAuthStore } from "@/stores/auth";
