+++ frontend/src/components/admin/drawer/admin-settings-button-icon.vue
