+++ frontend/src/components/settings/panel.vue
@@ -11,7 +11,7 @@ import { defineAsyncComponent, markRaw } from "vue";
 import { useAuthStore } from "@/stores/auth";
 
 const AdminMenu = markRaw(
-  defineAsyncComponent(() => import("@/components/admin/admin-menu.vue"))
+  defineAsyncComponent(() => import("@/components/admin/drawer/admin-menu.vue"))
 );
 import AuthMenu from "@/components/auth/auth-menu.vue";
